refactor(the structure of the project was restructed): restructing app core layers
The main idea is to follow the principles of Clean Architecture, and to advocate SRP for future
expansions.

BREAKING CHANGE: The project has been redesigned and there are now 3 main layers handler, service,
domain. Handler - exists to process incoming (currently http requests) and transfer them to the
service layer, which contains the basic business logic and calls to repo for interaction with
domain.